requirements.txt
fastapi
uvicorn
pydantic
pandas
python-jose[cryptography]
passlib[bcrypt]
pytest
requests